[Bhushan] Refactor. Get all employees updated

+++ Repository/Repo/NotesRepository/NotesRepo.cs
@@ -48,12 +48,21 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="notes">parameter notes</param>
         /// <returns>return notes</returns>
-        public async Task<Notes> AddNotes(Notes notes)
+        public async Task<Notes> AddNotes(Notes notes, string userEmail)
         {
             try
             {
-                this.context.Notes.Add(notes);
-                var result = this.context.SaveChangesAsync();
+                var val = this.context.Notes.Add(notes);
+                var result = await this.context.SaveChangesAsync();
+                Collaborater coll = new Collaborater
+                {
+                    NotesId = val.Entity.NotesId,
+                    UserID = notes.UserID,
+                    ReceiverEmail = userEmail,
+                    SenderEmail = userEmail
+                };
+                var res = this.context.Collaborater.Add(coll);
+                var results = this.context.SaveChangesAsync();
                 return await Task.Run(() => notes);
             }
             catch
@@ -67,12 +76,12 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">parameter ID</param>
         /// <returns>returns Notes</returns>
-        public async Task<Notes> DeleteNotes(int id)
+        public async Task<Notes> DeleteNotes(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
-                if (note != null && !note.IsTrash)
+                var note = this.GetNoteById(noteId).Result;
+                if (note != null && !note.IsTrash && note.UserID == userId)
                 {
                     note.IsTrash = true;
                     var result = this.context.SaveChangesAsync();
@@ -92,12 +101,12 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="notes">parameter notes</param>
         /// <returns>returns notes</returns>
-        public async Task<Notes> UpdateNotes(Notes notes)
+        public async Task<Notes> UpdateNotes(Notes notes, int userId)
         {
             try
             {
-                var note = this.context.Notes.Where(notesId => notesId.NotesId == notes.NotesId).SingleOrDefault();
-                if (note != null && !note.IsTrash && !note.IsArchive)
+                var note = this.context.Notes.Where(notesId => notesId.NotesId == userId).SingleOrDefault();
+                if (note != null && !note.IsTrash && !note.IsArchive && note.UserID == userId)
                 {
                     if (notes.Title != null)
                     {
@@ -131,13 +140,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">parameter ID</param>
         /// <returns>Returns Notes</returns>
-        public async Task<Notes> DeleteFromTrash(int id)
+        public async Task<Notes> DeleteFromTrash(int id, int userId)
         {
             try
             {
                 var note = this.GetNoteById(id).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null && note.IsTrash)
+                if (note != null && note.IsTrash && note.UserID == userId)
                 {
                     this.context.Notes.Remove(note);
                     var result = this.context.SaveChangesAsync();
@@ -153,52 +162,41 @@ namespace FundooRepository.Repo.NotesRepository
         }
 
         /// <summary>
-        /// Get all notes
-        /// </summary>
-        /// <returns>returns all notes</returns>
-        public async Task<IEnumerable<Notes>> GetAllNotes()
-        {
-            try
-            {
-                var allNotes = this.context.Notes.ToList();
-                if (allNotes != null)
-                {
-                    return await Task.Run(() => allNotes);
-                }
-
-                return null;
-            }
-            catch
-            {
-                throw new Exception();
-            }
-        }
-
-        /// <summary>
         /// Get all notes by email address
         /// </summary>
         /// <param name="email">parameter email</param>
         /// <returns>returns list of notes</returns>
-        public async Task<IEnumerable<Notes>> GetAllNotesByEmail(string email)
+        public async Task<List<Notes>> GetAllNotes(int userId)
         {
             try
             {
-                var allNotes = this.context.Notes.ToList();
                 List<Notes> list = new List<Notes>();
-                if (allNotes != null)
-                {
-                    foreach (Notes notes in allNotes)
-                    {
-                        if (notes.Email.Equals(email))
+                var allNotes = this.context.Notes.Join(this.context.Collaborater,
+                        Notes => Notes.NotesId,
+                        Collaborater => Collaborater.NotesId,
+                        (Notes, Collaborater) => new Notes
                         {
-                            list.Add(notes);
+                            NotesId = Collaborater.NotesId,
+                            UserID = Collaborater.UserID,
+                            Title = Notes.Title,
+                            Description = Notes.Description,
+                            CreatedTime = Notes.CreatedTime,
+                            ModifiedTime = Notes.ModifiedTime,
+                            Remainder = Notes.Remainder,
+                            Image = Notes.Image,
+                            IsArchive = Notes.IsArchive,
+                            IsPin = Notes.IsPin,
+                            IsTrash = Notes.IsTrash
                         }
+                       );
+                foreach (var data in allNotes)
+                {
+                    if (data.UserID == userId)
+                    {
+                        list.Add(data);
                     }
-
-                    return await Task.Run(() => list);
                 }
-
-                return null;
+                return list;
             }
             catch
             {
@@ -206,18 +204,26 @@ namespace FundooRepository.Repo.NotesRepository
             }
         }
 
+        public string GetUserIdByEmail(string email)
+        {
+            var account = this.context.Users.Where(accounts => accounts.Email == email).SingleOrDefault();
+            if (account == null)
+            {
+                return null;
+            }
+            return account.UserID.ToString();
+        }
         /// <summary>
         /// Set to zero in is Trash column
         /// </summary>
         /// <param name="id">Note ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> ResetIsTrash(int id)
+        public async Task<bool> ResetIsTrash(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
-                //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                var note = this.GetNoteById(noteId).Result;
+                if (note != null && note.UserID == userId)
                 {
                     note.IsTrash = false;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -237,13 +243,12 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Note ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> SetIsTrash(int id)
+        public async Task<bool> SetIsTrash(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
-                //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                var note = this.GetNoteById(noteId).Result;
+                if (note != null && note.UserID == userId)
                 {
                     note.IsTrash = true;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -263,13 +268,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Note ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> ResetArchive(int id)
+        public async Task<bool> ResetArchive(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
+                var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     note.IsArchive = false;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -289,13 +294,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Notes ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> SetArchive(int id)
+        public async Task<bool> SetArchive(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
+                var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     note.IsArchive = true;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -315,13 +320,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Notes Id</param>
         /// <returns>returns status</returns>
-        public async Task<bool> ResetPin(int id)
+        public async Task<bool> ResetPin(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
+                var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     note.IsPin = false;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -341,13 +346,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Notes ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> SetPin(int id)
+        public async Task<bool> SetPin(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
+                var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     note.IsPin = true;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -368,13 +373,13 @@ namespace FundooRepository.Repo.NotesRepository
         /// <param name="id">Notes ID</param>
         /// <param name="time">parameter time</param>
         /// <returns>returns status</returns>
-        public async Task<bool> AddRemainder(int id, string time)
+        public async Task<bool> AddRemainder(int noteId, string time, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
+                var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     note.Remainder = time;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -394,13 +399,12 @@ namespace FundooRepository.Repo.NotesRepository
         /// </summary>
         /// <param name="id">Notes ID</param>
         /// <returns>returns status</returns>
-        public async Task<bool> DeleteRemainder(int id)
+        public async Task<bool> DeleteRemainder(int noteId, int userId)
         {
             try
             {
-                var note = this.GetNoteById(id).Result;
-                //var note = this.context.Notes.Where(notesId => notesId.NotesId == id).SingleOrDefault();
-                if (note != null)
+                var note = this.GetNoteById(noteId).Result;
+                if (note != null && note.UserID == userId)
                 {
                     note.Remainder = null;
                     await Task.Run(() => this.context.SaveChangesAsync());
@@ -424,21 +428,28 @@ namespace FundooRepository.Repo.NotesRepository
         {
             try
             {
-                var note = this.GetNoteById(collaborater.NotesId).Result;
-                //var note = this.context.Notes.Where(noteId => noteId.NotesId == collaborater.NotesId).SingleOrDefault();
-                if (note != null)
+                string collaboratorUserId = this.GetUserIdByEmail(collaborater.ReceiverEmail);
+                int collUserId = int.Parse(collaboratorUserId);
+                bool duplicateStatus = this.CollaboratorDuplicatesCheck(collUserId, collaborater.NotesId, collaborater.ReceiverEmail);
+                if (collaboratorUserId != null && !duplicateStatus)
                 {
-                    Collaborater newCollaborater = new Collaborater
+                    var note = this.GetNoteById(collaborater.NotesId).Result;
+
+                    if (note != null && note.UserID == collaborater.UserID)
                     {
-                        NotesId = collaborater.NotesId,
-                        SenderEmail = collaborater.SenderEmail,
-                        ReceiverEmail = collaborater.ReceiverEmail
-                    };
-                    this.context.Collaboraters.Add(newCollaborater);
-                    var result = this.context.SaveChangesAsync();
-                    if (result != null)
-                    {
-                        return await Task.Run(() => collaborater);
+                        Collaborater newCollaborater = new Collaborater
+                        {
+                            SenderEmail = collaborater.SenderEmail,
+                            ReceiverEmail = collaborater.ReceiverEmail,
+                            UserID = collUserId,
+                            NotesId = collaborater.NotesId
+                        };
+                        this.context.Collaborater.Add(newCollaborater);
+                        var result = this.context.SaveChangesAsync();
+                        if (result != null)
+                        {
+                            return await Task.Run(() => collaborater);
+                        }
                     }
                 }
 
@@ -459,14 +470,14 @@ namespace FundooRepository.Repo.NotesRepository
         {
             try
             {
-                var collabrate = this.context.Collaboraters.Where(notes => notes.NotesId == collaborater.NotesId).ToList();
+                var collabrate = this.context.Collaborater.Where(notes => notes.NotesId == collaborater.NotesId).ToList();
                 if (collabrate != null)
                 {
                     foreach (Collaborater list in collabrate)
                     {
-                        if (list.ReceiverEmail == collaborater.ReceiverEmail)
+                        if (list.UserID == collaborater.UserID)
                         {
-                            this.context.Collaboraters.Remove(list);
+                            this.context.Collaborater.Remove(list);
                             await Task.Run(() => this.context.SaveChangesAsync());
                             return collaborater;
                         }
@@ -487,7 +498,7 @@ namespace FundooRepository.Repo.NotesRepository
         /// <param name="noteId">parameter note ID</param>
         /// <param name="imagePath">parameter image path</param>
         /// <returns>returns image upload result</returns>
-        public async Task<ImageUploadResult> UploadImage(int noteId, string imagePath)
+        public async Task<ImageUploadResult> UploadImage(int noteId, string imagePath, int userId)
         {
             try
             {
@@ -496,7 +507,7 @@ namespace FundooRepository.Repo.NotesRepository
                 string apiSecret = this.config["Cloudinary:APISecret"];
                 var note = this.GetNoteById(noteId).Result;
                 //var note = this.context.Notes.Where(noteAtId => noteAtId.NotesId == noteId).SingleOrDefault();
-                if (note != null)
+                if (note != null && note.UserID == userId)
                 {
                     Account account = new Account(cloudName, apiKey, apiSecret);
                     Cloudinary cloudinary = new Cloudinary(account);
@@ -519,18 +530,30 @@ namespace FundooRepository.Repo.NotesRepository
                 throw new Exception();
             }
         }
-        public async Task<Notes> GetNoteById(int id)
+        public async Task<Notes> GetNoteById(int noteId)
         {
             try
             {
-                var resultNote = this.GetNoteById(id).Result;
-                //var resultNote = this.context.Notes.Where(note => note.NotesId == id).SingleOrDefault();
+                var resultNote = this.context.Notes.Where(note => note.NotesId == noteId).SingleOrDefault();
                 return await Task.Run(() => resultNote);
             }
             catch
             {
                 throw new Exception();
             }
+        }
+
+        public bool CollaboratorDuplicatesCheck(int userId, int noteId, string receiverEmail)
+        {
+            var collaborate = this.context.Collaborater.Where(user => user.UserID == userId).ToList();
+            foreach (var collUsers in  collaborate)
+            {
+                if (collUsers.UserID == userId && collUsers.NotesId == noteId && collUsers.ReceiverEmail == receiverEmail)
+                {
+                    return true;
+                }
+            }
+            return false;
         }
     }
 }
